% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRWHN.R
\name{calculateRWHN}
\alias{calculateRWHN}
\alias{calculateTransitionMatrix}
\alias{getRWHN}
\alias{heatmap_RWHN}
\title{Calculate RWHN}
\usage{
calculateRWHN(
  hetNet,
  seeds,
  transitionProb,
  restart,
  eta_xy,
  eta_yz,
  eps = 1/10^6,
  random = F
)

calculateTransitionMatrix(transitionProb, hetNet, vertices)

getRWHN(transMat, restart, eta_xy, eta_yz, seeds, eps)

heatmap_RWHN(
  rwhn_output,
  database = "",
  colours = c(low = "#e6e4f8", high = "#46009e"),
  removeCommon = T
)
}
\arguments{
\item{hetNet}{List; output of constructHetNet()}

\item{seeds}{Character vector of seed nodes}

\item{transitionProb}{Integer; transition probability}

\item{restart}{Integer; restart probability}

\item{eta_xy}{Integer; weighting on protein layer}

\item{eta_yz}{Integer; weighting on function layer
\strong{For visualisation:}}

\item{eps}{Integer; steady state defined by distance between p0 and p1}

\item{random}{Logical; randomly permute edges? For control cases}

\item{vertices}{Data frame; vertices in heterogeneous network, as containined in output of constructHetNet()}

\item{transMat}{List; output of calculateTransitionMatrix()}

\item{rwhn_output}{output of calculateRWHN() (or getRWHN())}

\item{database}{character; name of annotation database}

\item{colours}{character vector of length 2 with colours for scale}

\item{removeCommon}{logical; if rwhn_output is a list, remove functions ranked at the same position in all cases. This is useful for removing noise.}
}
\value{
A data frame of ranked nodes from the functional layer
}
\description{
Calculate RWHN
}
