% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrepresentationAnalysis.R
\name{overrepresentationAnalysis}
\alias{overrepresentationAnalysis}
\title{Overrepresentation analysis}
\usage{
overrepresentationAnalysis(
  clustering,
  database = "GO_Biological_Process_2018",
  simplify = T,
  vis = F,
  colours = c(low = "#439981", high = "#b2d4cd"),
  RWHN_sig = NULL
)
}
\arguments{
\item{clustering}{A named vector; names correspond to sites in format \emph{Gene name}_\emph{AminoAcid+Position} (e.g. MAPK1_T185), values correspond to cluster. This is}

\item{database}{Annotation database to calculate enrichment from. See \href{https://maayanlab.cloud/Enrichr/#stats}{enrichR website}}

\item{simplify}{If database = "GO_Biological_Process_2018", remove redundant terms}

\item{vis}{logical; visualise output?}

\item{colours}{if vis = T, a character vector of length 2 for colour scale}

\item{RWHN_sig}{optional; results of RWHN if wanting to perform comparison,}
}
\value{
If vis = T, a ggplot object. If vis = F, a data frame containing enriched terms.
}
\description{
Overrepresentation analysis
}
