% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichr.R
\name{getEnrichr}
\alias{getEnrichr}
\title{Get EnrichR results from API}
\usage{
getEnrichr(genes, enrichrLib, pval)
}
\arguments{
\item{genes}{character vector of HGNC symbols}

\item{enrichrLib}{EnrichR library to use;}

\item{pval}{enrichment pvalue cut off}
}
\value{
dataframe; enrichR results
}
\description{
see https://maayanlab.cloud/Enrichr/#stats
for EnrichR libraries
}
