% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyGO.R
\name{simplifyGO}
\alias{simplifyGO}
\title{Remove semantically similar GO terms.
Based on code by Jamie Soul:
https://github.com/soulj/SkeletalVis-Pipeline/blob/master/src/SYBIL\%20Systems\%20Biology/GOEnrichment.R}
\usage{
simplifyGO(GOID, semData)
}
\arguments{
\item{GOID}{List of GOID's to simplify}

\item{semData}{output from GOSemSim::godata}
}
\value{
vector of reduced GO terms
}
\description{
Remove semantically similar GO terms.
Based on code by Jamie Soul:
https://github.com/soulj/SkeletalVis-Pipeline/blob/master/src/SYBIL\%20Systems\%20Biology/GOEnrichment.R
}
